+++ Fyber/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
+* 7.8.6.0
+  * This version of the adapters has been certified with Fyber Marketplace 7.8.6 and MoPub SDK 5.17.0.
+  * Do not pass GDPR consent to Fyber via adapters.
+
 * 7.8.5.0 
   * Initial Commit
   * Add support for banner, interstitial and rewarded video.